refactor: moved cheese enum to be inner class of BaseCheeseWheel
