+++ src/main/java/growthcraft/milk/block/CheeseWheelAgeableBlock.java
@@ -1,7 +1,5 @@
 package growthcraft.milk.block;
 
-import growthcraft.milk.datagen.Cheese;
-
 public class CheeseWheelAgeableBlock extends BaseCheeseWheel {
     public CheeseWheelAgeableBlock(Cheese variant) {
         super(variant);
