+++ src/main/java/growthcraft/milk/block/CheeseWheelProcessedBlock.java
@@ -1,12 +1,9 @@
 package growthcraft.milk.block;
 
-import growthcraft.milk.block.entity.CheeseWheelBlockEntity;
-import growthcraft.milk.datagen.Cheese;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.Items;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.BlockHitResult;
