+++ src/main/java/growthcraft/milk/block/CheeseWheelWaxableBlock.java
@@ -1,6 +1,5 @@
 package growthcraft.milk.block;
 
-import growthcraft.milk.datagen.Cheese;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
